+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/BrowserToolbar.kt
@@ -14,12 +14,14 @@ import android.widget.ImageButton
 import androidx.annotation.DrawableRes
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.VisibleForTesting.PRIVATE
+import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.core.view.forEach
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancelChildren
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
+import mozilla.components.browser.toolbar.behavior.BrowserToolbarBehavior
 import mozilla.components.browser.toolbar.display.DisplayToolbar
 import mozilla.components.browser.toolbar.edit.EditToolbar
 import mozilla.components.concept.toolbar.AutocompleteDelegate
@@ -317,6 +319,20 @@ class BrowserToolbar @JvmOverloads constructor(
      */
     override fun dismissMenu() {
         display.views.menu.dismissMenu()
+    }
+
+    override fun enableScrolling() {
+        // Behavior can be changed without us knowing. Not safe to use a memoized value.
+        (layoutParams as? CoordinatorLayout.LayoutParams)?.apply {
+            (behavior as? BrowserToolbarBehavior)?.enableScrolling()
+        }
+    }
+
+    override fun disableScrolling() {
+        // Behavior can be changed without us knowing. Not safe to use a memoized value.
+        (layoutParams as? CoordinatorLayout.LayoutParams)?.apply {
+            (behavior as? BrowserToolbarBehavior)?.disableScrolling()
+        }
     }
 
     internal fun onUrlEntered(url: String) {