+++ mobile/android/android-components/components/feature/customtabs/src/test/java/mozilla/components/feature/customtabs/feature/CustomTabSessionTitleObserverTest.kt
@@ -92,6 +92,8 @@ class CustomTabSessionTitleObserverTest {
         override fun displayMode() = Unit
         override fun editMode() = Unit
         override fun dismissMenu() = Unit
+        override fun enableScrolling() = Unit
+        override fun disableScrolling() = Unit
     }
 }