+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -113,6 +113,14 @@ class ToolbarAutocompleteFeatureTest {
         override fun dismissMenu() {
             fail()
         }
+
+        override fun enableScrolling() {
+            fail()
+        }
+
+        override fun disableScrolling() {
+            fail()
+        }
     }
 
     @Test